Move ContentType into BtpSubprotocol
