+++ btp-core/src/main/java/org/interledger/btp/BtpSubProtocol.java
@@ -20,11 +20,18 @@ package org.interledger.btp;
  * =========================LICENSE_END==================================
  */
 
+import static java.lang.String.format;
+
 import org.interledger.annotations.Immutable;
+import org.interledger.core.InterledgerPreparePacket;
+import org.interledger.encoding.asn.framework.CodecContext;
 
 import org.immutables.value.Value.Default;
 
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.util.Objects;
 
 /**
  * Contains information required to process sub-protocols using BTP.
@@ -43,8 +50,8 @@ public interface BtpSubProtocol {
   /**
    * The content-type of this sub-protocol.
    */
-  default BtpSubProtocolContentType getContentType() {
-    return BtpSubProtocolContentType.MIME_APPLICATION_OCTET_STREAM;
+  default ContentType getContentType() {
+    return ContentType.MIME_APPLICATION_OCTET_STREAM;
   }
 
   /**
@@ -59,14 +66,57 @@ public interface BtpSubProtocol {
   }
 
   /**
+   * A content-type descriptor for BTP sub-protocol payloads, mainly used for logging and smart deserializing.
+   */
+  enum ContentType {
+    MIME_APPLICATION_OCTET_STREAM((short) 0),
+    MIME_TEXT_PLAIN_UTF8((short) 1),
+    MIME_APPLICATION_JSON((short) 2),
+    ;
+
+    private final short code;
+
+
+    ContentType(short code) {
+      this.code = code;
+    }
+
+    /**
+     * Get a new {@link ContentType} from the given code.
+     *
+     * @param code a type code
+     *
+     * @return an instance of {@link ContentType}
+     */
+    public static ContentType fromCode(short code) {
+
+      switch (code) {
+        case 0:
+          return MIME_APPLICATION_OCTET_STREAM;
+        case 1:
+          return MIME_TEXT_PLAIN_UTF8;
+        case 2:
+          return MIME_APPLICATION_JSON;
+        default:
+          throw new IllegalArgumentException(
+              format("Unknown BTP Sub-Protocol Content Type: %s", code));
+      }
+    }
+
+    public short getCode() {
+      return this.code;
+    }
+  }
+
+  /**
    * Abstract implementation to satisfy the Immutables library.
    */
   @Immutable
   abstract class AbstractBtpSubProtocol implements BtpSubProtocol {
 
     @Default
-    public BtpSubProtocolContentType getContentType() {
-      return BtpSubProtocolContentType.MIME_APPLICATION_OCTET_STREAM;
+    public ContentType getContentType() {
+      return ContentType.MIME_APPLICATION_OCTET_STREAM;
     }
 
     @Default
