+++ btp-core/src/test/java/org/interledger/btp/BtpMessageTest.java
@@ -9,9 +9,9 @@ package org.interledger.btp;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,12 +20,13 @@ package org.interledger.btp;
  * =========================LICENSE_END==================================
  */
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+
+import org.interledger.btp.BtpSubProtocol.ContentType;
 
 import org.junit.Test;
 
 import java.nio.charset.StandardCharsets;
-import java.time.Instant;
 
 public class BtpMessageTest {
 
@@ -33,7 +34,7 @@ public class BtpMessageTest {
   private static final BtpSubProtocols SUB_PROTOCOLS = new BtpSubProtocols();
   private static final BtpSubProtocol SUB_PROTOCOL = BtpSubProtocol.builder()
       .protocolName("TEST")
-      .contentType(BtpSubProtocolContentType.MIME_TEXT_PLAIN_UTF8)
+      .contentType(ContentType.MIME_TEXT_PLAIN_UTF8)
       .data("Test Data".getBytes(StandardCharsets.UTF_8))
       .build();
