+++ btp-core/src/test/java/org/interledger/btp/BtpResponseTest.java
@@ -22,6 +22,8 @@ package org.interledger.btp;
 
 import static org.junit.Assert.assertEquals;
 
+import org.interledger.btp.BtpSubProtocol.ContentType;
+
 import org.junit.Test;
 
 import java.nio.charset.StandardCharsets;
@@ -32,7 +34,7 @@ public class BtpResponseTest {
   private static final BtpSubProtocols SUB_PROTOCOLS = new BtpSubProtocols();
   private static final BtpSubProtocol SUB_PROTOCOL = BtpSubProtocol.builder()
       .protocolName("TEST")
-      .contentType(BtpSubProtocolContentType.MIME_TEXT_PLAIN_UTF8)
+      .contentType(ContentType.MIME_TEXT_PLAIN_UTF8)
       .data("Test Data".getBytes(StandardCharsets.UTF_8))
       .build();
