+++ btp-core/src/test/java/org/interledger/btp/BtpRuntimeExceptionTest.java
@@ -24,6 +24,8 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertEquals;
 
+import org.interledger.btp.BtpSubProtocol.ContentType;
+
 import org.junit.Test;
 
 import java.nio.charset.StandardCharsets;
@@ -40,7 +42,7 @@ public class BtpRuntimeExceptionTest {
   static {
     SUB_PROTOCOLS.add(BtpSubProtocol.builder()
         .protocolName("TEST")
-        .contentType(BtpSubProtocolContentType.MIME_TEXT_PLAIN_UTF8)
+        .contentType(ContentType.MIME_TEXT_PLAIN_UTF8)
         .data("Test Data".getBytes(StandardCharsets.UTF_8))
         .build());
   }
